Add core game loop, undo and JSON validation
The game loop is centered around the endpoint

    POST /nim/:id/take

which allows a user to take 1, 2, or 3 matchsticks. The server responds
with the result of that turn as well as the computer's response.

If at any time you're not happy with how the turn went you can undo your actions by calling

    POST /nim/:id/undo